restuct ProfileView
