+++ MealBytes/View/Profile/ProfileView/ProfileViewModel/PasswordError.swift
@@ -1,0 +1,22 @@
+//
+//  PasswordError.swift
+//  MealBytes
+//
+//  Created by Vlad Posherstnik on 07/08/2025.
+//
+
+import SwiftUI
+
+enum PasswordError: Error {
+    case tooShort
+    case mismatch
+    
+    var message: String {
+        switch self {
+        case .tooShort:
+            return ProfileMessage.passwordTooShort.text
+        case .mismatch:
+            return ProfileMessage.passwordMismatch.text
+        }
+    }
+}
